Remove the range slider and display bar graphs
The range slider was not a great control for exploring the data, didn't
work well on mobile devices, and required third party dependencies.

This commit displays a summary per year of the measurements beside the
map. It's not a perfect solution but it does make better use of the
screen space and provides an rough idea of the cleanliness of the river
over time at a quick glance.

While the measurement level colours work well on the map, they don't
stand out enough in this visual summary.