Add possible solution for displaying about content
It would be nice to use as little JavaScript as possible and
particularly avoid it for display text that the user should be able to
see without JavaScript. However, the side panel where the bar graphs are
displayed is an ideal place to display the about text, as opposed to
sending the user to a completely different page.

I will go with this solution for now and revisit in due course.